requirements.txt
# Versiones probadas y compatibles con Python 3.9
streamlit>=1.30,<1.38
pandas>=2.0,<2.2
numpy>=1.24,<2.0
plotly>=5.17,<6.0
python-dotenv>=1.0
openpyxl>=3.1
xlsxwriter>=3.1
PyYAML>=6.0
bcrypt>=4.0
